refactor trace cid handling
