+++ contract_review_app/core/trace.py
@@ -1,7 +1,11 @@
 from __future__ import annotations
 
 from collections import OrderedDict
+import hashlib
+import json
 from typing import Any, Dict
+
+from fastapi import Request
 
 
 class TraceStore:
@@ -26,3 +30,29 @@ class TraceStore:
 
     def list(self) -> list[str]:
         return list(self._data.keys())
+
+
+def compute_cid(request: Request) -> str:
+    """Compute deterministic content id for a request.
+
+    POST/PUT/PATCH: path + sorted(query) + canonical JSON body
+    GET/DELETE: path + sorted(query)
+    """
+    path = request.url.path
+    query_items = sorted(request.query_params.multi_items())
+    query = "&".join(f"{k}={v}" for k, v in query_items)
+    body_part = ""
+    if request.method.upper() in {"POST", "PUT", "PATCH"}:
+        body_bytes = getattr(request.state, "body", b"")
+        try:
+            obj: Any = json.loads(body_bytes.decode("utf-8")) if body_bytes else None
+        except Exception:
+            obj = body_bytes.decode("utf-8", "ignore")
+        if obj is None:
+            body_part = ""
+        elif isinstance(obj, (dict, list)):
+            body_part = json.dumps(obj, sort_keys=True, ensure_ascii=False)
+        else:
+            body_part = str(obj)
+    raw = f"{path}{query}{body_part}".encode("utf-8")
+    return hashlib.sha256(raw).hexdigest()
